HIVE-28904: Remove Arrow from Hive
