+++ pom.xml
@@ -109,8 +109,6 @@
     <!-- Make sure to sync it with standalone-metastore/pom.xml -->
     <antlr4.version>4.9.3</antlr4.version>
     <apache-directory-server.version>2.0.0-M24</apache-directory-server.version>
-    <!-- Include arrow for LlapOutputFormatService -->
-    <arrow.version>16.0.0</arrow.version>
     <avatica.version>1.12.0</avatica.version>
     <avro.version>1.11.4</avro.version>
     <bcprov-jdk18on.version>1.78</bcprov-jdk18on.version>
