+++ ql/src/java/org/apache/hadoop/hive/ql/optimizer/physical/Vectorizer.java
@@ -71,7 +71,6 @@ import org.apache.hadoop.hive.ql.exec.mr.MapRedTask;
 import org.apache.hadoop.hive.ql.exec.persistence.MapJoinKey;
 import org.apache.hadoop.hive.ql.exec.tez.TezTask;
 import org.apache.hadoop.hive.ql.exec.vector.VectorExpressionDescriptor;
-import org.apache.hadoop.hive.ql.exec.vector.filesink.VectorFileSinkArrowOperator;
 import org.apache.hadoop.hive.ql.exec.vector.mapjoin.VectorMapJoinInnerBigOnlyLongOperator;
 import org.apache.hadoop.hive.ql.exec.vector.mapjoin.VectorMapJoinInnerBigOnlyMultiKeyOperator;
 import org.apache.hadoop.hive.ql.exec.vector.mapjoin.VectorMapJoinInnerBigOnlyStringOperator;
@@ -4319,48 +4318,6 @@ public class Vectorizer implements PhysicalPlanResolver {
     return true;
   }
 
-  private boolean checkForArrowFileSink(FileSinkDesc fileSinkDesc,
-      boolean isTez, VectorizationContext vContext,
-      VectorFileSinkDesc vectorDesc) throws HiveException {
-
-    // Various restrictions.
-
-    boolean isVectorizationFileSinkArrowNativeEnabled =
-        HiveConf.getBoolVar(hiveConf,
-            HiveConf.ConfVars.HIVE_VECTORIZATION_FILESINK_ARROW_NATIVE_ENABLED);
-
-    String engine = HiveConf.getVar(hiveConf, HiveConf.ConfVars.HIVE_EXECUTION_ENGINE);
-
-    String serdeClassName = fileSinkDesc.getTableInfo().getSerdeClassName();
-
-    boolean isOkArrowFileSink =
-        serdeClassName.equals("org.apache.hadoop.hive.ql.io.arrow.ArrowColumnarBatchSerDe") &&
-        isVectorizationFileSinkArrowNativeEnabled &&
-        engine.equalsIgnoreCase("tez");
-
-    return isOkArrowFileSink;
-  }
-
-  private Operator<? extends OperatorDesc> specializeArrowFileSinkOperator(
-      Operator<? extends OperatorDesc> op, VectorizationContext vContext, FileSinkDesc desc,
-      VectorFileSinkDesc vectorDesc) throws HiveException {
-
-    Class<? extends Operator<?>> opClass = VectorFileSinkArrowOperator.class;
-
-    Operator<? extends OperatorDesc> vectorOp = null;
-    try {
-      vectorOp = OperatorFactory.getVectorOperator(
-          opClass, op.getCompilationOpContext(), op.getConf(),
-          vContext, vectorDesc);
-    } catch (Exception e) {
-      LOG.info("Vectorizer vectorizeOperator file sink class exception " + opClass.getSimpleName() +
-          " exception " + e);
-      throw new HiveException(e);
-    }
-
-    return vectorOp;
-  }
-
   private boolean usesVectorUDFAdaptor(VectorExpression vecExpr) {
     if (vecExpr == null) {
       return false;
@@ -5278,7 +5235,7 @@ public class Vectorizer implements PhysicalPlanResolver {
     // This "global" allows various validation methods to set the "not vectorized" reason.
     currentOperator = op;
 
-    boolean isNative;
+    boolean isNative = false;
     try {
       switch (op.getType()) {
         case MAPJOIN:
@@ -5326,7 +5283,6 @@ public class Vectorizer implements PhysicalPlanResolver {
                 vectorOp = OperatorFactory.getVectorOperator(
                     opClass, op.getCompilationOpContext(), desc,
                     vContext, vectorMapJoinDesc);
-                isNative = false;
               } else {
 
                 // TEMPORARY Until Native Vector Map Join with Hybrid passes tests...
@@ -5490,20 +5446,11 @@ public class Vectorizer implements PhysicalPlanResolver {
             FileSinkDesc fileSinkDesc = (FileSinkDesc) op.getConf();
 
             VectorFileSinkDesc vectorFileSinkDesc = new VectorFileSinkDesc();
-            boolean isArrowSpecialization =
-                checkForArrowFileSink(fileSinkDesc, isTez, vContext, vectorFileSinkDesc);
 
-            if (isArrowSpecialization) {
-              vectorOp =
-                  specializeArrowFileSinkOperator(
-                      op, vContext, fileSinkDesc, vectorFileSinkDesc);
-              isNative = true;
-            } else {
-              vectorOp =
-                  OperatorFactory.getVectorOperator(
-                      op.getCompilationOpContext(), fileSinkDesc, vContext, vectorFileSinkDesc);
-              isNative = false;
-            }
+            vectorOp =
+                OperatorFactory.getVectorOperator(
+                    op.getCompilationOpContext(), fileSinkDesc, vContext, vectorFileSinkDesc);
+            isNative = false;
           }
           break;
         case LIMIT:
